LayoutTests/inspector/debugger/script-formatter.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="resources/obfuscated.js"></script>

<script>

var test = function()
{
    var worker = new Worker("ScriptFormatterWorker.js");

    InspectorTest.runDebuggerTestSuite([
        function testScriptFormatterWorker(next)
        {
            worker.onmessage = InspectorTest.safeWrap(function(event)
            {
                InspectorTest.assertEquals("var x = 0\n", event.data.content);
                next();
            });

            worker.onerror = function(event)
            {
                InspectorTest.addResult("Error in worker: " + event.data);
                next();
            };

            worker.postMessage({ mimeType: "text/javascript", content: "var x=0" });
        },

        function testSourceMapping(next)
        {
            worker.onmessage = InspectorTest.safeWrap(function(event)
            {
                var source = WebInspector.panels.scripts.visibleView._content;
                var formattedSource = event.data.content;
                var mapping = event.data.mapping;

                function testMapping(string)
                {
                    var position = source.indexOf(string);
                    var index = mapping.original.upperBound(position) - 1;
                    var delta = position - mapping.original[index];
                    var formattedPosition = Math.min(mapping.formatted[index] + delta, mapping.formatted[index + 1]);
                    InspectorTest.assertEquals(string, formattedSource.substr(formattedPosition, string.length));
                }

                testMapping("onmessage");
                testMapping("indent_start");
                testMapping("function require");
                testMapping("var regexp");
                testMapping("importScripts");

                next();
            });

            worker.onerror = function(event)
            {
                InspectorTest.addResult("Error in worker: " + event.data);
                next();
            };

            InspectorTest.showScriptSource("obfuscated.js", didShowScriptSource);

            function didShowScriptSource(sourceFrame)
            {
                worker.postMessage({ mimeType: "text/javascript", content: sourceFrame._content });
            }
        },

        function testFormatInlinedScripts(next)
        {
            worker.onmessage = InspectorTest.safeWrap(function(event)
            {
                InspectorTest.addResult(event.data.content);
                next();
            });

            worker.onerror = function(event)
            {
                InspectorTest.addResult("Error in worker: " + event.data);
                next();
            };

            var content = "<html><body><script>function f(){}<" + "/script><script>function g(){}<" + "/script></body></html>";
            worker.postMessage({ mimeType: "text/html", content: content });
        }
    ]);
}

</script>

</head>

<body onload="runTest()">
<p>Tests the script formatting functionality.
</p>

</body>
</html>
