LayoutTests/inspector/debugger/script-formatter-expected.txt

Tests the script formatting functionality.

Debugger was enabled.

Running: testScriptFormatterWorker

Running: testSourceMapping

Running: testFormatInlinedScripts
<html><body><script>
function f() {
}
</script><script>
function g() {
}
</script></body></html>
Debugger was disabled.
